+++ src/script/config.js
@@ -122,7 +122,7 @@ export var config = {
 \\caption{Quicksort}
 \\begin{algorithmic}
 \\PROCEDURE{Quicksort}{$A, p, r$}
-    \\IF{$p < r$} 
+    \\IF{$p < r$}
         \\STATE $q = $ \\CALL{Partition}{$A, p, r$}
         \\STATE \\CALL{Quicksort}{$A, p, q - 1$}
         \\STATE \\CALL{Quicksort}{$A, q + 1, r$}